Rename to chat text.
